requirements.txt
urllib3==1.23
distro==1.3.0
pandas==0.23.4
tabula-py==1.3.0
